added delete functionality to comments

+++ src/components/comments/PostComments.js
@@ -2,6 +2,7 @@ import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { getCommentsByPost } from "../../managers/comments";
 import { getPostById } from "../../managers/posts";
+import { getCurrentAuthor } from "../../managers/users";
 
 export const PostComments = ({ token }) => {
   const { postId } = useParams();
@@ -19,24 +20,31 @@ export const PostComments = ({ token }) => {
     }
   }, [postId]);
 
-  const deleteButton = (comment) => {
-    return (
-      <button
-        onClick={() => {
-          fetch(`http://localhost:8088/comments/${comment.id}`, {
-            method: "DELETE",
-          }).then(() => {
-            getCommentsByPost(postId).then((foundComments) =>
-              setComments(foundComments)
-            );
-          });
-        }}
-        className="submission__delete small-button"
-      >
-        Delete
-      </button>
+  const [currentAuthorId, setCurrentAuthorId] = useState({});
+
+  const deleteButton = (comment, event) => {
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this comment?"
     );
+    if (confirmDelete) {
+      return fetch(`http://localhost:8000/comments/${comment.id}`, {
+        method: "DELETE",
+        headers: {
+          authorization: `Token ${localStorage.getItem("auth_token")}`,
+        },
+      }).then(() => {
+        getCommentsByPost(postId).then((foundComments) =>
+          setComments(foundComments)
+        );
+      });
+    }
   };
+
+  useEffect(() => {
+    getCurrentAuthor(token).then((author) => {
+      setCurrentAuthorId(author);
+    });
+  }, [token]);
 
   const sortedComments = comments.sort((a, b) => {
     return new Date(b.created_at) - new Date(a.created_at);
@@ -54,11 +62,22 @@ export const PostComments = ({ token }) => {
               User: {comment.author?.first_name} {comment.author?.last_name}
             </div>
             <div>Created: {comment.date_created}</div>
-            {comment.author_id === parseInt(token) ? (
-              deleteButton(comment)
-            ) : (
-              <div></div>
-            )}
+            <div>
+              {console.log("Comment Author ID:", comment.author?.id)}
+              {console.log("Current Author ID:", currentAuthorId)}
+
+              {comment.author?.id === currentAuthorId[0].id ? (
+                <button
+                  onClick={(event) => {
+                    deleteButton(comment, event);
+                  }}
+                >
+                  Delete
+                </button>
+              ) : (
+                ""
+              )}
+            </div>
           </section>
         );
       })}
